requirements.txt
absl-py==0.10.0
albumentations==0.4.6
antlr4-python3-runtime==4.8
attrs==20.1.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
comet-git-pure==0.19.16
comet-ml==3.2.2
configobj==5.0.6
configparser==5.0.0
cycler==0.10.0
decorator==4.4.2
docker-pycreds==0.4.0
dulwich==0.20.6
everett==1.0.2
filelock==3.0.12
future==0.18.2
gdown==3.12.2
gitdb==4.0.5
GitPython==3.1.7
google-auth==1.20.1
google-auth-oauthlib==0.4.1
gql==0.2.0
graphql-core==1.1
grpcio==1.31.0
Hydra==2.5
hydra-core==1.0.0rc4
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-resources==3.0.0
joblib==0.16.0
jsonschema==3.2.0
kiwisolver==1.2.0
kornia==0.5.8
Markdown==3.2.2
matplotlib==3.3.1
munch==2.5.0
netifaces==0.10.9
networkx==2.5
numpy==1.19.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
omegaconf==2.0.1rc12
opencv-python==4.4.0.42
opencv-python-headless==4.4.0.42
packaging==20.4
pandas==1.1.1
pathtools==0.1.2
Pillow==7.2.0
pretrainedmodels==0.7.4
promise==2.3
protobuf==3.13.0
psutil==5.7.2
ptflops==0.6.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-lightning==0.9.0
pytorchcv==0.0.58
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
sentry-sdk==0.17.0
Shapely==1.7.1
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
socks==0
subprocess32==3.5.4
tensorboard==2.2.0
tensorboard-plugin-wit==1.7.0
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2020.8.13
timm==0.2.1
torch==1.9.0
torchinfo==1.5.3
torchstat==0.0.7
torchsummary==1.5.1
torchvision==0.10.0
tqdm==4.48.2
typing-extensions==3.7.4.3
urllib3==1.25.10
wandb==0.10.2
watchdog==0.10.3
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
wurlitzer==2.0.1
zipp==3.5.0
